update_server/requirements.txt
Flask==0.9
PyYAML==3.10
